==> ALDEx2/06_totalRNA_DEG_analysis.ipynb <==
2 main objectives:

1) join mixo orf annotations to DEGs.
2) count frequencies of species.

discussed polyA / totalRNA differences in annotation in notebook \#1.
Analysis so far has been done on totalRNA.

`ls -lh /tscc/projects/ps-allenlab/archdata/zfussy/RAP_runs/NCOG_totalRNA/ | grep zfussy` to separate out recent files from Dr. Venepally's (much older) base files.

`ls -lh /tscc/projects/ps-allenlab/archdata/zfussy/RAP_runs/NCOG_totalRNA/ | grep :` to separate out files edited this year (will have a HH:MM after their "last date modified" instead of a year)


"upreg" = upreg in group B (higher than median temp shallow sample) than group A (lower than median)...

**Joining mixo ORF annotations to DEGs**

import pandas as pd

#file too big. have to read it line by line.
annotation_columns = pd.read_csv("nov2024_totalRNA_RAPresults/NCOG-totalRNA-annotations.tsv", sep = "\t", index_col=0,
                                  nrows=1, header = None).values.tolist()[0]

len(annotation_columns)

upreg_DEGs = pd.read_csv("05_out_upreg_pval.tsv", sep = "\t", index_col = 0)
downreg_DEGs = pd.read_csv("05_out_downreg_pval.tsv", sep = "\t", index_col = 0)

upreg_DEGs[list(annotation_columns)] = ""
downreg_DEGs[list(annotation_columns)] = ""

#large file, read line by line
for line in open("nov2024_totalRNA_RAPresults/NCOG-totalRNA-annotations.tsv", "r"):
    terms = line.strip("\n").split("\t")
    assert(len(terms) == 19)
    contig_id = terms[0]
    annotations = terms[1:]
    if contig_id in upreg_DEGs.index:
        upreg_DEGs.loc[contig_id, annotation_columns] = annotations
    elif contig_id in downreg_DEGs.index:
        downreg_DEGs.loc[contig_id, annotation_columns] = annotations

upreg_DEGs

with pd.option_context("display.max_colwidth", None):
    display(downreg_DEGs["LPI_taxonomy"])

**Binning LPI taxonomy with search_species**

#upreg_orf_taxonomies = upreg_DEGs["LPI_taxonomy"].apply(search_species)
#downreg_orf_taxonomies = downreg_DEGs["LPI_taxonomy"].apply(search_species)
upreg_orf_taxonomies = upreg_DEGs["LPI_taxonomy"].apply(lambda x: x.strip(";.").split(";")[-1])
downreg_orf_taxonomies = downreg_DEGs["LPI_taxonomy"].apply(lambda x: x.strip(";.").split(";")[-1])

#upregulated in higher than "median temp samples", relative to "lower than median temp" samples.
upreg_orf_taxonomies.value_counts(dropna=False)

downreg_orf_taxonomies.value_counts(dropna=False)

**Other frequency tables**

annotation_columns

upreg_DEGs.loc[upreg_orf_taxonomies.str.contains("Emiliania"), "GOterms"].value_counts()

downreg_DEGs.loc[downreg_orf_taxonomies.str.contains("Emiliania"), "KO_description"].value_counts()

upreg_DEGs["KO_description"].value_counts()

**Emiliania specific analysis**

I might need to bring back effect size to make sure 1) these aren't just all the Emiliania genes in the original contig set 2) that they aren't "sig downreg" because a few 1s were 0s instead. Can load the original frequencies from file for a post-facto check... or can see if effect size serves as a proxy for this.

**"Determines the effect size, which is the median of the ratio of the between-group difference and the
larger of the variance within groups."** from [docs](https://www.bioconductor.org/packages/devel/bioc/manuals/ALDEx2/man/ALDEx2.pdf).
- if this isn't 1, means that within group variance (within the group with larger spread) is larger than the variance between groups... this isn't exactly how an ANOVA works. there, it's that the sum of the within-group variances has to be larger than the all-data variance. Is this an arbitrary threshold, should I consider scale effects?
- what is the "median of the ratio"? for 2 values, isn't the median just the average anyways? Which 2 values?

#the suggested cutoff for effect size is 1... none of these meet that.
downreg_DEGs.loc[downreg_orf_taxonomies.str.contains("Emiliania"), ["effect", "Pfam_description"]]

annotation_columns

downreg_DEGs.loc[downreg_orf_taxonomies.str.contains("Emiliania"), "KO_description"].value_counts(sort=True)

downreg_DEGs.loc[downreg_orf_taxonomies.str.contains("Emiliania"), "Pfam_description"].value_counts(sort=True)

downreg_DEGs.loc[downreg_orf_taxonomies.str.contains("Emiliania"), "EG5_Description"].value_counts(sort=True)

upreg_DEGs["NCBIfam_description"].value_counts(sort=True)

downreg_DEGs["NCBIfam_description"].value_counts(sort=True)

upreg_DEGs["EG5_Description"].value_counts(sort=True)

downreg_DEGs["EG5_Description"].value_counts(sort=True)
